Created checkout folder and moved files there to better organize components
