+++ customer_portal/src/components/checkout/Confirmation.jsx
