+++ customer_portal/src/components/checkout/OrderItem.jsx
