+++ customer_portal/src/components/checkout/OrderSummary.jsx
